flash_attention/pyproject.toml
[project]
name = "flash-attention"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torchvision>=0.26.0",
]
